requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
deepface==0.0.93
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Session==0.8.0
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-api-core==2.10.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
gunicorn==20.1.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgspec==0.19.0
mtcnn==0.1.1
namex==0.0.9
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==1.5.3
pillow==11.2.1
protobuf==4.25.7
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
tqdm==4.67.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0